+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configurations/SettingsTests.cs
@@ -252,6 +252,7 @@ public class SettingsTests : IDisposable
     [InlineData("ELASTICTRANSPORT", TracerInstrumentation.ElasticTransport)]
     [InlineData("KAFKA", TracerInstrumentation.Kafka)]
     [InlineData("ORACLEMDA", TracerInstrumentation.OracleMda)]
+    [InlineData("RABBITMQ6", TracerInstrumentation.RabbitMq6)]
     internal void TracerSettings_Instrumentations_SupportedValues(string tracerInstrumentation, TracerInstrumentation expectedTracerInstrumentation)
     {
         Environment.SetEnvironmentVariable(ConfigurationKeys.Traces.TracesInstrumentationEnabled, "false");